requirements.txt
Django==1.11.2
django-filter
mysqlclient==1.3.10
social-auth-app-django
django-extensions
Pillow
djangorestframework==3.8.2

